refactor(credit): modularize ic_critic_engine → critic/ package (Wave 1 PR #1)
Converts the monolithic ic_critic_engine.py (766 LOC) into an edgar-style
package with strict DAG: models.py (leaf) → domain modules → service.py.

Package structure:
  critic/
    models.py        — CriticVerdict (frozen, slots), INSTRUMENT_TYPE_PROFILES
    classifier.py    — classify_instrument_type (public, was _classify_instrument_type)
    macro_checks.py  — deterministic macro-consistency flags
    prompt_builder.py — build_critic_prompt, build_critic_input
    parser.py        — parse_critic_response, clamp
    service.py       — critique_intelligence, build_critic_packet (orchestrator)
    __init__.py      — standard imports, __all__

Changes:
- Never-raises contract: service.py catches all exceptions, returns NOT_ASSESSED
- CriticVerdict now frozen=True, slots=True (reconstructed instead of mutated)
- stdlib logging → structlog
- _classify_instrument_type → classify_instrument_type (public API)
- All callers updated: deep_review.py, deep_review_prompts.py
- test_vertical_engines.py EXPECTED_MODULES updated
- Old flat file deleted
